requirements.txt
# Core dependencies
pandas>=2.1.0
numpy>=1.24.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.20
alembic>=1.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=23.1.0
click>=8.1.7
schedule>=1.2.0

# Trading & Market Data
alpaca-py>=0.25.0
ib-insync>=0.9.86
requests>=2.31.0
yfinance>=0.2.27

# Data Analysis & Backtesting
scipy>=1.11.0
scikit-learn>=1.3.0
numba>=0.58.0

# Dashboard & Visualization
streamlit>=1.26.0
plotly>=5.15.0
matplotlib>=3.7.2
seaborn>=0.12.2

# Testing & Code Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5

# Production & Monitoring
gunicorn>=21.2.0
prometheus-client>=0.17.0
sentry-sdk>=1.29.0